Re-organize folders/files -- create a shared folder
